Start looking at the read side processing.
Queries should be hitting a read-side projection rather than the state of the application in most cases are the queries are likely to be more complex than grabbing identifiers. Queries should not be commands either (CQRS explains there is a separation between the two!), meaning that the read-side projection should always be hit. Once I have gotten Cassandra object mapping working (or just a set of queries which might be more performant!), we can ensure that the read-side processors create performant projections that we can use to power the GET portion of SOAD.

We do have the benefit that the GraphQL layer should adjust the projection further, so we may not need view models for some endpoints on the services, but we may need to consider it if we expose any endpoints as REST endpoints to be used by the public.