This was an interesting experiment but it shows us that we can't really do it this way.
Ths idea was to try to create a service that validates a token, though we'd want to pull the profile back. Unfortunately, this hasn't worked as intended. While I have no more time tonight, the next idea is to use topics. If not, then we'll be publishing events to set the keys and ensuring all services that need it have the methods they need (which will only be IP or JWT auth that needs spreading out)